+++ changelog.md
@@ -1,6 +1,7 @@
 # Jumpscare Changelog
 ## v1.3.2
 - Updated for 2.206 (Geode v3.0.0-beta.1)
+- Added a button to open the config folder in the mod's settings popup
 ## v1.3.1
 - Added an option to choose a specific jumpscare to be used
 - Added an option to show the mod setting buttons in the pause menu